Sensors
Attempted to create 3 sensors with the recommended specs:
* Include memory of observations from (forwards-directed) ray traces of previous
frames. Note that you must implement this in your own code, not through a simple
parameter change of existing code.
• Simulate sound by including observations of nearby (moving) objects. Note that, in
this case (without vision), it would likely be impossible to hear whether an observed
agent is a teammate or an opponent.
• Decouple the movement direction from the vision cone, simulating an agent that can
move in one direction whilst looking in a different direction. The vision angle may
either be an additional action of the agent, determined randomly or according to a
handcrafted heuristic, or move according to some regular pattern.

However, I cannot append them to the agents in Unity (even though Unity sees the code files). Will ask the TA about this and will update the PDF with documentation on everything I have done and tried.